Refactor Detalles
