+++ DREAM Lab Frontend/src/pages/HomePage/components/Detalles/Detalles.css
